refactor(common): move Uniq to common package to reduce import cycles
Move the generic Uniq function from utils to common package since common
has fewer dependencies and is lower in the import hierarchy. Also update
HasSheets to use common.Uniq instead of slices.Sort + slices.Compact.
