+++ backend/common/util.go
@@ -232,3 +232,19 @@ func IsNil(val any) bool {
 func NewP[T any](x T) *T {
 	return &x
 }
+
+// Uniq returns a new slice with duplicate elements removed, preserving order.
+func Uniq[T comparable](array []T) []T {
+	res := make([]T, 0, len(array))
+	seen := make(map[T]struct{}, len(array))
+
+	for _, e := range array {
+		if _, ok := seen[e]; ok {
+			continue
+		}
+		seen[e] = struct{}{}
+		res = append(res, e)
+	}
+
+	return res
+}
