+++ backend/utils/utils.go
@@ -197,21 +197,6 @@ func CheckDatabaseGroupMatch(ctx context.Context, expression string, database *s
 	return false, nil
 }
 
-func Uniq[T comparable](array []T) []T {
-	res := make([]T, 0, len(array))
-	seen := make(map[T]struct{}, len(array))
-
-	for _, e := range array {
-		if _, ok := seen[e]; ok {
-			continue
-		}
-		seen[e] = struct{}{}
-		res = append(res, e)
-	}
-
-	return res
-}
-
 // IsSpaceOrSemicolon checks if the rune is a space or a semicolon.
 func IsSpaceOrSemicolon(r rune) bool {
 	if ok := unicode.IsSpace(r); ok {
